requirements.txt
fastapi==0.109.0
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.1.0
uvicorn==0.25.0
PyJWT==2.8.0